requirements.txt
python-dotenv==1.0.0
requests==2.28.2
termcolor==2.2.0
tqdm==4.65.0